+++ src/fields/CategoriesSearch.php
@@ -15,6 +15,7 @@ use spicyweb\oddsandends\assetbundles\tools\ToolsAsset;
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @author Supercool
  * @since 2.0.0
+ * @deprecated in 5.1.0; use \craft\fields\Categories with showSearchInput = true instead
  */
 class CategoriesSearch extends Categories
 {